+++ scripts/build_app.py
@@ -73,8 +73,8 @@ APP_RESOURCES = (
     ("Resources/AppIcon.icns", "AppIcon.icns"),
     ("Resources/Assets.car", "Assets.car"),
     (
-        "Sources/ArknightsClient/Resources/GameIconBackground.png",
-        "GameIconBackground.png",
+        "Sources/ArknightsClient/Resources/OperatorIconFrame.svg",
+        "OperatorIconFrame.svg",
     ),
 )